requirements.txt
spotipy==2.23.0
python-dotenv==1.0.0
seaborn
plotly
yellowbrick
flask-cors
concurrently --save-dev